+++ CHANGELOG.md
@@ -41,7 +41,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* [9526](https://github.com/osmosis-labs/osmosis/pull/9526) feat: inflation endpoint
 * [9527](https://github.com/osmosis-labs/osmosis/pull/9527) unfork cometbft
+
 
 ## v30.0.3